Made a new src directory: "core"
For code that implements the core of the analyzer, rather than shoving everything into analyzer.ts or the toplevel src/ directory.
