+++ src/analyzer.ts
@@ -14,31 +14,13 @@
 
 /// <reference path="../custom_typings/main.d.ts" />
 
-import * as path from 'path';
-
-import {AnalysisCache} from './analysis-cache';
-import {CssParser} from './css/css-parser';
-import {HtmlCustomElementReferenceScanner} from './html/html-element-reference-scanner';
-import {HtmlImportScanner} from './html/html-import-scanner';
-import {HtmlParser} from './html/html-parser';
-import {HtmlScriptScanner} from './html/html-script-scanner';
-import {HtmlStyleScanner} from './html/html-style-scanner';
-import {JavaScriptParser} from './javascript/javascript-parser';
-import {JsonParser} from './json/json-parser';
-import {Document, InlineDocInfo, LocationOffset, ScannedDocument, ScannedElement, ScannedFeature, ScannedImport, ScannedInlineDocument} from './model/model';
-import {ParsedDocument} from './parser/document';
+import {AnalyzerCacheContext} from './core/analyzer-cache-context';
+import {Document} from './model/model';
 import {Parser} from './parser/parser';
-import {Measurement, TelemetryTracker} from './perf/telemetry';
-import {BehaviorScanner} from './polymer/behavior-scanner';
-import {CssImportScanner} from './polymer/css-import-scanner';
-import {DomModuleScanner} from './polymer/dom-module-scanner';
-import {PolymerElementScanner} from './polymer/polymer-element-scanner';
-import {scan} from './scanning/scan';
+import {Measurement} from './perf/telemetry';
 import {Scanner} from './scanning/scanner';
 import {UrlLoader} from './url-loader/url-loader';
 import {UrlResolver} from './url-loader/url-resolver';
-import {ElementScanner as VanillaElementScanner} from './vanilla-custom-elements/element-scanner';
-import {Severity, Warning, WarningCarryingException} from './warning/warning';
 
 export interface Options {
   urlLoader: UrlLoader;
@@ -115,447 +97,5 @@ export class Analyzer {
    */
   async load(resolvedUrl: string, providedContents?: string) {
     return this._cacheContext.load(resolvedUrl, providedContents);
-  }
-}
-
-/**
- * Represents an Analyzer with a given AnalysisCache instance.
- *
- * Used to provide a consistent cache in the face of updates happening in
- * parallel with analysis work. A given AnalyzerCacheContext is forked via
- * either the fileChanged or clearCaches methods.
- *
- * For almost all purposes this is an entirely internal implementation detail.
- */
-export class AnalyzerCacheContext {
-  private _parsers = new Map<string, Parser<ParsedDocument<any, any>>>([
-    ['html', new HtmlParser()],
-    ['js', new JavaScriptParser({sourceType: 'script'})],
-    ['css', new CssParser()],
-    ['json', new JsonParser()],
-  ]);
-
-  /** A map from import url to urls that document lazily depends on. */
-  private _lazyEdges: LazyEdgeMap|undefined;
-
-  private _scanners: ScannerTable;
-
-  private _loader: UrlLoader;
-  private _resolver: UrlResolver|undefined;
-
-  private _cache = new AnalysisCache();
-
-  private _telemetryTracker = new TelemetryTracker();
-  private _generation = 0;
-
-  private static _getDefaultScanners(lazyEdges: LazyEdgeMap|undefined) {
-    return new Map<string, Scanner<any, any, any>[]>([
-      [
-        'html',
-        [
-          new HtmlImportScanner(lazyEdges),
-          new HtmlScriptScanner(),
-          new HtmlStyleScanner(),
-          new DomModuleScanner(),
-          new CssImportScanner(),
-          new HtmlCustomElementReferenceScanner()
-        ]
-      ],
-      [
-        'js',
-        [
-          new PolymerElementScanner(),
-          new BehaviorScanner(),
-          new VanillaElementScanner()
-        ]
-      ],
-    ]);
-  }
-
-  constructor(options: Options) {
-    this._loader = options.urlLoader;
-    this._resolver = options.urlResolver;
-    this._parsers = options.parsers || this._parsers;
-    this._lazyEdges = options.lazyEdges;
-    this._scanners = options.scanners ||
-        AnalyzerCacheContext._getDefaultScanners(this._lazyEdges);
-  }
-
-  /**
-   * Returns a copy of this cache context with proper cache invalidation.
-   */
-  filesChanged(urls: string[]) {
-    const newCache =
-        this._cache.invalidatePaths(urls.map(url => this._resolveUrl(url)));
-    return this._fork(newCache);
-  }
-
-
-  /**
-   * Implements Analyzer#analyze, see its docs.
-   */
-  async analyze(url: string, contents?: string): Promise<Document> {
-    const resolvedUrl = this._resolveUrl(url);
-
-    const cachedResult = this._cache.analyzedDocumentPromises.get(resolvedUrl);
-    if (cachedResult) {
-      return cachedResult;
-    }
-
-    const promise = (async() => {
-      // Make sure we wait and return a Promise before doing any work, so that
-      // the Promise is cached before anything else happens.
-      await Promise.resolve();
-      const doneTiming =
-          this._telemetryTracker.start('analyze: make document', url);
-      const scannedDocument = await this._scan(resolvedUrl, contents);
-      if (scannedDocument === 'visited') {
-        throw new Error(
-            `This should not happen. Got a cycle of length zero(!) scanning ${url
-            }`);
-      }
-      const document = this._makeDocument(scannedDocument);
-      doneTiming();
-      return document;
-    })();
-    this._cache.analyzedDocumentPromises.set(resolvedUrl, promise);
-    return promise;
-  }
-
-  /**
-   * Constructs a new analyzed Document and adds it to the analyzed Document
-   * cache.
-   */
-  private _makeDocument(scannedDocument: ScannedDocument): Document {
-    const resolvedUrl = scannedDocument.url;
-
-    if (this._cache.analyzedDocuments.has(resolvedUrl)) {
-      throw new Error(`Internal error: document ${resolvedUrl} already exists`);
-    }
-
-    const document = new Document(scannedDocument, this);
-    if (!this._cache.analyzedDocumentPromises.has(resolvedUrl)) {
-      this._cache.analyzedDocumentPromises.set(
-          resolvedUrl, Promise.resolve(document));
-    }
-    this._cache.analyzedDocuments.set(resolvedUrl, document);
-    document.resolve();
-    return document;
-  }
-
-  /**
-   * Gets an analyzed Document from the document cache. This is only useful for
-   * Analyzer plugins. You almost certainly want to use `analyze()` instead.
-   *
-   * If a document has been analyzed, it returns the analyzed Document. If not
-   * the scanned document cache is used and a new analyzed Document is returned.
-   * If a file is in neither cache, it returns `undefined`.
-   */
-  _getDocument(url: string): Document|undefined {
-    const resolvedUrl = this._resolveUrl(url);
-    let document = this._cache.analyzedDocuments.get(resolvedUrl);
-    if (document) {
-      return document;
-    }
-    const scannedDocument = this._cache.scannedDocuments.get(resolvedUrl);
-    return scannedDocument && this._makeDocument(scannedDocument);
-  }
-
-  /**
-   * This is only useful for Analyzer plugins.
-   *
-   * If a url has been scanned, returns the ScannedDocument.
-   */
-  _getScannedDocument(url: string): ScannedDocument|undefined {
-    const resolvedUrl = this._resolveUrl(url);
-    return this._cache.scannedDocuments.get(resolvedUrl);
-  }
-
-  async getTelemetryMeasurements(): Promise<Measurement[]> {
-    return this._telemetryTracker.getMeasurements();
-  }
-
-  /**
-   * Clear all cached information from this analyzer instance.
-   *
-   * Note: if at all possible, instead tell the analyzer about the specific
-   * files that changed rather than clearing caches like this. Caching provides
-   * large performance gains.
-   */
-  clearCaches(): AnalyzerCacheContext {
-    return this._fork(new AnalysisCache());
-  }
-
-  /**
-   * Return a copy, but with the given cache.
-   */
-  private _fork(cache: AnalysisCache): AnalyzerCacheContext {
-    const copy = new AnalyzerCacheContext({
-      lazyEdges: this._lazyEdges,
-      parsers: this._parsers,
-      scanners: this._scanners,
-      urlLoader: this._loader,
-      urlResolver: this._resolver
-    });
-    copy._telemetryTracker = this._telemetryTracker;
-    copy._cache = cache;
-    copy._generation = this._generation + 1;
-    return copy;
-  }
-
-  /**
-   * Scan a toplevel document given its url and optionally its contents.
-   */
-  private async _scan(
-      resolvedUrl: string, contents?: string,
-      visited?: Set<string>): Promise<ScannedDocument|'visited'> {
-    if (visited && visited.has(resolvedUrl)) {
-      return 'visited';
-    }
-    visited = visited || new Set();
-    visited.add(resolvedUrl);
-    const cachedResult = this._cache.scannedDocumentPromises.get(resolvedUrl);
-    if (cachedResult) {
-      await this._scanDependenciesOfToplevelDoc(await cachedResult, visited);
-      return cachedResult;
-    }
-    const promise = (async() => {
-      // Make sure we wait and return a Promise before doing any work, so that
-      // the Promise is cached before anything else happens.
-      await Promise.resolve();
-      const document = await this._parse(resolvedUrl, contents);
-      return this._scanDocument(document, visited);
-    })();
-    this._cache.scannedDocumentPromises.set(resolvedUrl, promise);
-    const scannedDocument = await promise;
-    await this._scanDependenciesOfToplevelDoc(scannedDocument, visited);
-    return scannedDocument;
-  }
-
-  /**
-   * Parses and scans a document from source.
-   */
-  private async _scanInlineSource(
-      type: string, contents: string, url: string, visited: Set<string>,
-      inlineInfo?: InlineDocInfo<any>,
-      attachedComment?: string): Promise<ScannedDocument> {
-    const resolvedUrl = this._resolveUrl(url);
-    const parsedDoc =
-        this._parseContents(type, contents, resolvedUrl, inlineInfo);
-    const scannedDoc =
-        await this._scanDocument(parsedDoc, visited, attachedComment);
-    return scannedDoc;
-  }
-
-  /**
-   * Scans a ParsedDocument.
-   */
-  private async _scanDocument(
-      document: ParsedDocument<any, any>, visited: Set<string>,
-      maybeAttachedComment?: string): Promise<ScannedDocument> {
-    const warnings: Warning[] = [];
-    const scannedFeatures = await this._getScannedFeatures(document);
-    // If there's an HTML comment that applies to this document then we assume
-    // that it applies to the first feature.
-    const firstScannedFeature = scannedFeatures[0];
-    if (firstScannedFeature && firstScannedFeature instanceof ScannedElement) {
-      firstScannedFeature.applyHtmlComment(maybeAttachedComment);
-    }
-
-    const scannedDocument =
-        new ScannedDocument(document, scannedFeatures, warnings);
-
-    if (!scannedDocument.isInline) {
-      if (this._cache.scannedDocuments.has(scannedDocument.url)) {
-        throw new Error(
-            'Scanned document already in cache. This should never happen.');
-      }
-      this._cache.scannedDocuments.set(scannedDocument.url, scannedDocument);
-    }
-    await this._scanInlineDocuments(scannedDocument, visited);
-
-    return scannedDocument;
-  }
-
-  private async _scanDependenciesOfToplevelDoc(
-      scannedDocument: ScannedDocument, visited: Set<string>) {
-    const scanDepPromise =
-        this._cache.dependenciesScanned.get(scannedDocument.url);
-    if (scanDepPromise) {
-      return scanDepPromise;
-    }
-    const promise = (async() => {
-      await Promise.resolve();
-      await this._scanImports(scannedDocument, visited);
-    })();
-    this._cache.dependenciesScanned.set(scannedDocument.url, promise);
-    return promise;
-  }
-
-  private async _scanImports(
-      scannedDocument: ScannedDocument, visited: Set<string>) {
-    const scannedImports = scannedDocument.getNestedFeatures().filter(
-        (e) => e instanceof ScannedImport) as ScannedImport[];
-    for (const scannedImport of scannedImports) {
-      // TODO(garlicnation): Move this logic into model/document. During
-      // the recursive feature walk, features from lazy imports
-      // should be marked.
-      if (scannedImport.type !== 'lazy-html-import') {
-        await this._scanImport(
-            scannedImport, scannedDocument.warnings, visited);
-      }
-    }
-
-    this._cache.dependencyGraph.addDependenciesOf(
-        scannedDocument.url,
-        scannedImports.map(imp => this._resolveUrl(imp.url)));
-  }
-
-  private async _scanInlineDocuments(
-      containingDocument: ScannedDocument, visited: Set<string>) {
-    for (const feature of containingDocument.features) {
-      if (!(feature instanceof ScannedInlineDocument)) {
-        continue;
-      }
-      await this._scanInlineDocument(
-          feature,
-          containingDocument.url,
-          containingDocument.warnings,
-          visited);
-    }
-  }
-
-  /**
-   * Scan an inline document found within a containing parsed doc.
-   */
-  private async _scanInlineDocument(
-      inlineDoc: ScannedInlineDocument, url: string, warnings: Warning[],
-      visited: Set<string>): Promise<ScannedDocument|null> {
-    const locationOffset: LocationOffset = {
-      line: inlineDoc.locationOffset.line,
-      col: inlineDoc.locationOffset.col,
-      filename: url
-    };
-    const inlineInfo = {locationOffset, astNode: inlineDoc.astNode};
-    try {
-      const scannedDocument = await this._scanInlineSource(
-          inlineDoc.type,
-          inlineDoc.contents,
-          url,
-          visited,
-          inlineInfo,
-          inlineDoc.attachedComment);
-      inlineDoc.scannedDocument = scannedDocument;
-      return scannedDocument;
-    } catch (err) {
-      if (err instanceof WarningCarryingException) {
-        warnings.push(err.warning);
-        return null;
-      }
-      throw err;
-    }
-  }
-
-  private async _scanImport(
-      scannedImport: ScannedImport, warnings: Warning[],
-      visited: Set<string>): Promise<null> {
-    const url = this._resolveUrl(scannedImport.url);
-    try {
-      await this._scan(url, undefined, visited);
-    } catch (error) {
-      if (error instanceof NoKnownParserError) {
-        // We probably don't want to fail when importing something
-        // that we don't know about here.
-        return null;
-      }
-      error = error || '';
-      warnings.push({
-        code: 'could-not-load',
-        message: `Unable to load import: ${error.message || error}`,
-        sourceRange:
-            (scannedImport.urlSourceRange || scannedImport.sourceRange)!,
-        severity: Severity.ERROR
-      });
-      return null;
-    }
-    return null;
-  }
-
-  /**
-   * Loads the content at the provided resolved URL.
-   *
-   * Currently does no caching. If the provided contents are given then they
-   * are used instead of hitting the UrlLoader (e.g. when you have in-memory
-   * contents that should override disk).
-   */
-  async load(resolvedUrl: string, providedContents?: string) {
-    if (!this._loader.canLoad(resolvedUrl)) {
-      throw new Error(`Can't load URL: ${resolvedUrl}`);
-    }
-    return providedContents == null ? await this._loader.load(resolvedUrl) :
-                                      providedContents;
-  }
-
-  private async _parse(resolvedUrl: string, providedContents?: string):
-      Promise<ParsedDocument<any, any>> {
-    const cachedResult = this._cache.parsedDocumentPromises.get(resolvedUrl);
-    if (cachedResult) {
-      return cachedResult;
-    }
-
-    // Use an immediately executed async function to create the final Promise
-    // synchronously so we can store it in this._documents before any other
-    // async operations to avoid any race conditions.
-    const promise = (async() => {
-      // Make sure we wait and return a Promise before doing any work, so that
-      // the Promise can be cached.
-      await Promise.resolve();
-
-      const content = await this.load(resolvedUrl, providedContents);
-      const extension = path.extname(resolvedUrl).substring(1);
-
-      const doneTiming = this._telemetryTracker.start('parse', 'resolvedUrl');
-      const parsedDoc = this._parseContents(extension, content, resolvedUrl);
-      doneTiming();
-      return parsedDoc;
-    })();
-    this._cache.parsedDocumentPromises.set(resolvedUrl, promise);
-    return promise;
-  }
-
-  private _parseContents(
-      type: string, contents: string, url: string,
-      inlineInfo?: InlineDocInfo<any>): ParsedDocument<any, any> {
-    const parser = this._parsers.get(type);
-    if (parser == null) {
-      throw new NoKnownParserError(`No parser for for file type ${type}`);
-    }
-    try {
-      return parser.parse(contents, url, inlineInfo);
-    } catch (error) {
-      if (error instanceof WarningCarryingException) {
-        throw error;
-      }
-      throw new Error(`Error parsing ${url}:\n ${error.stack}`);
-    }
-  }
-
-  private async _getScannedFeatures(document: ParsedDocument<any, any>):
-      Promise<ScannedFeature[]> {
-    const scanners = this._scanners.get(document.type);
-    if (scanners) {
-      return scan(document, scanners);
-    }
-    return [];
-  }
-
-  /**
-   * Resolves a URL with this Analyzer's `UrlResolver` if it has one, otherwise
-   * returns the given URL.
-   */
-  private _resolveUrl(url: string): string {
-    return this._resolver && this._resolver.canResolve(url) ?
-        this._resolver.resolve(url) :
-        url;
   }
 }
