+++ src/core/dependency-graph.ts
